+++ src/main/java/com/example/duantotnghiepgiaythethaonova/service/GioHangChiTietService.java
@@ -5,13 +5,14 @@ package com.example.duantotnghiepgiaythethaonova.service;
 import com.example.duantotnghiepgiaythethaonova.dto.GioHangChiTietDTO;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 @Service
 public interface GioHangChiTietService  {
 
 	List<GioHangChiTietDTO> findAllByGioHangId(Integer id);
 
-	void capNhatSoLuongGioHangChiTiet(Integer[] ids, Integer[] soLuongs,Integer[] donGias);
+	void capNhatSoLuongGioHangChiTiet(Integer[] ids, Integer[] soLuongs, BigDecimal[] donGias);
 
 	void capNhatGioHangThanhDaXoaById(Integer id);
